+++ web/src/admin.ts
@@ -217,6 +217,8 @@ export function build_page(): void {
         web_stream_unreads_count_display_policy_values:
             settings_config.web_stream_unreads_count_display_policy_values,
         web_animate_image_previews_values: settings_config.web_animate_image_previews_values,
+        resolved_topic_notice_auto_read_policy_values:
+            settings_config.resolved_topic_notice_auto_read_policy_values,
         color_scheme_values: settings_config.color_scheme_values,
         web_home_view_values: settings_config.web_home_view_values,
         settings_object: realm_user_settings_defaults,