+++ zerver/lib/events.py
@@ -105,6 +105,7 @@ from zerver.models.realms import (
     get_realm_domains,
 )
 from zerver.models.streams import get_default_stream_groups
+from zerver.models.users import ResolvedTopicNoticeAutoReadPolicyEnum
 from zerver.tornado.django_api import get_user_events, request_event_queue
 from zproject.backends import email_auth_enabled, password_auth_enabled
 
@@ -595,6 +596,11 @@ def fetch_initial_state_data(
         state["realm_user_settings_defaults"]["available_notification_sounds"] = (
             get_available_notification_sounds()
         )
+        state["realm_user_settings_defaults"]["resolved_topic_notice_auto_read_policy"] = (
+            ResolvedTopicNoticeAutoReadPolicyEnum(
+                realm_user_default.resolved_topic_notice_auto_read_policy
+            ).name
+        )
 
     if want("realm_domains"):
         state["realm_domains"] = get_realm_domains(realm)
@@ -838,6 +844,11 @@ def fetch_initial_state_data(
         state["user_settings"]["timezone"] = canonicalize_timezone(settings_user.timezone)
         state["user_settings"]["available_notification_sounds"] = (
             get_available_notification_sounds()
+        )
+        state["user_settings"]["resolved_topic_notice_auto_read_policy"] = (
+            ResolvedTopicNoticeAutoReadPolicyEnum(
+                settings_user.resolved_topic_notice_auto_read_policy
+            ).name
         )
 
     if want("user_status"):